+++ agentops/llms/ollama.py
@@ -2,7 +2,9 @@ import inspect
 import sys
 from typing import Optional
 
-from agentops import LLMEvent, Session
+from ..event import ActionEvent, ErrorEvent, LLMEvent
+from ..session import Session
+from ..log_config import logger
 from agentops.helpers import get_ISO_time, check_call_stack_for_agent_id
 
 original_func = {}
